6.7: anecdotes, step5
