+++ redux-anecdotes/src/components/AnecdoteForm.js
@@ -1,0 +1,25 @@
+import { useDispatch } from "react-redux";
+import { newAnecdote } from "../reducers/anecdoteReducer";
+
+const AnecdoteForm = () => {
+  const dispatch = useDispatch();
+  
+  const create = (event) => {
+    event.preventDefault();
+    const anecdote = event.target.anecdote.value;
+    event.target.anecdote.value = '';
+    dispatch(newAnecdote(anecdote));
+  }
+  
+  return(
+    <div>
+      <h2>create new</h2>
+      <form onSubmit={create}>
+        <div><input name="anecdote"/></div>
+        <button type="submit">create</button>
+      </form>
+    </div>
+  )
+}
+
+export default AnecdoteForm;
